* fix-39: refactor ScenarioTest and PageObjects

* fix-39: add CloudModule

* fix-39: add PicEngine

* fix-39: add SharedPreferencesWrapper

* fix-39: rework UiState

* fix-39: add small fixes

* fix-39: Add CloudModule

* fix-39: Add PicEngine

* fix-39: Add SharedPreferencesWrapper

* fix-39: Rework UiState

* fix-39: add small fixes

* Merge remote-tracking branch 'origin/fix/39-refactoring-after-review'…

* fix-39: add MediaItem extension

* fix-39: mix refactor

* fix: refactor ui tests #39

* fix: refactor ui tests #38

* Merge remote-tracking branch 'origin/fix/39-refactoring-after-review'…

* Merge remote-tracking branch 'origin/fix/39-refactoring-after-review'…

* fix: rework UiObserver

* fix: update dependencies versions

* fix: add some documentation

* fix: add some documentation

* Merge remote-tracking branch 'origin/fix/39-refactoring-after-review'…

+++ app/src/main/java/com/ru/androidexperts/muzicapp/search/data/cache/CacheDataSource.kt
@@ -16,7 +16,7 @@ interface CacheDataSource {
             val termId = dao.saveTerm(TermsCache(term = term))
             val tracksIds = dao.saveTracks(tracks)
             tracksIds.forEach {
-                dao.saveTrackIdByTerm(TrackIdByTermCache(termId.toInt(), it))
+                dao.saveTrackIdByTerm(TrackIdByTermCache(termId, it))
             }
         }
